requirements.txt
dashscope
numpy
openwakeword
sounddevice
soundfile
langchain-core
langchain-openai
